Correct filename of task file
This fixes this problem:
[DEPRECATION WARNING]: Included file 'install-apt-dependencies.yml' not found, however since this
include is not explicitly marked as 'static: yes', we will try and include it
dynamically later. In the future, this will be an error unless 'static: no' is
used on the include task. If you do not want missing includes to be considered
dynamic, use 'static: yes' on the include or set the global ansible.cfg options
 to make all inclues static for tasks and/or handlers.
This feature will be
removed in a future release. Deprecation warnings can be disabled by setting
deprecation_warnings=False in ansible.cfg.